cosmosdb-history/requirements.txt
langchain
openai
pypdf
streamlit
python-dotenv
azure-cosmos